pyproject.toml
[tool.poetry]
name = "pochurl-rim"
version = "0.1.0"
description = "Pochurl - Read It Later!"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
